+++ provider/resource_project.go
@@ -907,21 +907,22 @@ func resourceProjectUpdate(ctx context.Context, d *schema.ResourceData, meta int
 	}
 
 	var maintenanceWindow = new(neon.MaintenanceWindow)
-	if v, ok := d.GetOk("maintenance_window"); ok {
-		if len(v.([]interface{})) > 0 && v.([]interface{})[0] != nil {
-			if v, ok := v.([]interface{})[0].(map[string]interface{}); ok && len(v) > 0 {
-				weekdaysRaw := v["weekdays"].([]interface{})
-				var weekdays = make([]int, len(weekdaysRaw))
-				for i, weekday := range weekdaysRaw {
-					weekdays[i] = weekday.(int)
+	maintenanceWindow = nil
+	if d.HasChange("maintenance_window") {
+		if v, ok := d.GetOk("maintenance_window"); ok {
+			if len(v.([]interface{})) > 0 && v.([]interface{})[0] != nil {
+				if v, ok := v.([]interface{})[0].(map[string]interface{}); ok && len(v) > 0 {
+					weekdaysRaw := v["weekdays"].([]interface{})
+					var weekdays = make([]int, len(weekdaysRaw))
+					for i, weekday := range weekdaysRaw {
+						weekdays[i] = weekday.(int)
+					}
+					maintenanceWindow.Weekdays = weekdays
+					maintenanceWindow.StartTime = v["start_time"].(string)
+					maintenanceWindow.EndTime = v["end_time"].(string)
 				}
-				maintenanceWindow.Weekdays = weekdays
-				maintenanceWindow.StartTime = v["start_time"].(string)
-				maintenanceWindow.EndTime = v["end_time"].(string)
 			}
 		}
-	} else {
-		maintenanceWindow = nil
 	}
 
 	withLogicalReplication := types.GetTristateBool(d, "enable_logical_replication")