+++ CHANGELOG.md
@@ -13,7 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- [[218](https://github.com/kislerdm/terraform-provider-neon/issues/218)] Fixed state management for the `enable_logical_replication` of the resource `neon_project`. It can be changed in-place now.
+- [[218](https://github.com/kislerdm/terraform-provider-neon/issues/218)] Fixed state management for the attribute `enable_logical_replication` of the resource `neon_project`. It can be changed in-place now.
+- [[223](https://github.com/kislerdm/terraform-provider-neon/issues/223)] Fixed state management for the attribute `maintenance_window` of the resource `neon_project`. It will be updated only if it's changed in the tf state change.
 
 ## [v0.13.0] - 2026-01-02